2015 update
This is a summary and update of what has been added or changed in the last year. The code hasn't gotten any prettier and relies a lot on the implementation structure I've created around it, but it should be pretty easy to hack together if given a new set of fields feeding into it.